// apps/web/src/api/generated/schemas/ticketCreate.ts
/**
 * Generated by orval v7.15.0 🍺
 * Do not edit manually.
 * x-ear API
 * Backend API for x-ear CRM
 * OpenAPI spec version: 1.0.0
 */
import type { TicketCreateDescription } from './ticketCreateDescription';
import type { TicketPriority } from './ticketPriority';
import type { TicketCategory } from './ticketCategory';
import type { TicketCreateTenantId } from './ticketCreateTenantId';

export interface TicketCreate {
  title: string;
  description?: TicketCreateDescription;
  priority?: TicketPriority;
  category?: TicketCategory;
  tenantId?: TicketCreateTenantId;
}