// apps/web/src/api/generated/schemas/ticketCategory.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * x-ear API
 * Backend API for x-ear CRM
 * OpenAPI spec version: 1.0.0
 */

export type TicketCategory = typeof TicketCategory[keyof typeof TicketCategory];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const TicketCategory = {
  general: 'general',
  technical: 'technical',
  billing: 'billing',
  feature_request: 'feature_request',
} as const;