// apps/web/src/api/generated/schemas/ticketPriority.ts
/**
 * Generated by orval v7.15.0 🍺
 * Do not edit manually.
 * x-ear API
 * Backend API for x-ear CRM
 * OpenAPI spec version: 1.0.0
 */

export type TicketPriority = typeof TicketPriority[keyof typeof TicketPriority];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const TicketPriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  urgent: 'urgent',
} as const;